requirements.txt
Django>=5.1.3
djangorestframework>=3.15.2
pillow>=11.0.0
django-cors-headers>=4.6.0
tensorflow>=2.17.0
keras>=3.4.1
numpy>=1.26.4
matplotlib>=3.8.0
pandas>=2.2.2
